Making the line up stop when lined up
Explained: a dummy variable ‘a’ is true, which is hooked up to a loop
that will (hopefully) line the robot up on the X and Y axises. When it
is lined up for 1 second, it will set a to false, and will break from
the loop, so it ends the program.